[Client|API] rename service to non-Angular std
